Simple Crosshair
So I made a super simple crosshair, just take this png, make a new material in unity, drop this png onto that material. In the inspector for the material go to shader > Gui > Text Shader. Then in the hierarchy > Create > UI > Image. Then Drop the new cross hair material into the material slot and make sure the position is 0, 0, 0. This will create a crosshair that follows the players view, directly in the center of the screen.