Super Simple Beams
The plus and minus are for a particle effect. I have it working on my end I just have to figure out how to export particle effects from Unity. 
For now attach the beams to the shrink ray and enter these values:
Position - X: -0.0199 Y: 0.0049 Z: 0.1255
Rotation - X: -2.569 Y: -11.669 Z: -4.735
Scale - X: 0.02994886 Y: 0.02994888 Z:  0.02994887

Now if you guys could just add a quick activation script to this to make them turn on and off with the shrink and grow button presses that would be great. Also adding a quick rotation script to the beams would be great just to make them twist or rotate when visible and active would be great.