Particle Settings
Just take the meshes and materials of the plus and minus that I uploaded earlier and add them to these prefabs under the Renderer tab of the particle. It should hopefully work correctly then. If the materials for some reason didn't export with the meshes, just add a blue color mat to the minus and an orange color mat to the plus and you have basically the same thing. Then parent the particle effects to the beams and you should then have a working effect... hopefully.